added check for if environment exists, and separated out execute commands into separate files
