+++ LadybugTools_Adapter/AdapterActions/Execute.cs
@@ -69,586 +69,8 @@ namespace BH.Adapter.LadybugTools
                 return new List<object>();
             }
 
+            //See .cs files in Execute folder for possible commands
             return RunCommand(command as dynamic, actionConfig);
-        }
-
-        /**************************************************/
-        /* Private methods - Run Ladybug Command          */
-        /**************************************************/
-
-        private List<object> RunCommand(GetMaterialCommand command, ActionConfig actionConfig)
-        {
-            LadybugConfig config;
-
-            if (actionConfig?.GetType() == typeof(LadybugConfig))
-            {
-                config = (LadybugConfig)actionConfig;
-                config.JsonFile = new FileSettings()
-                {
-                    FileName = $"LBTBHoM_Materials.json",
-                    Directory = Path.GetTempPath()
-                };
-            }
-            else
-            {
-                config = new LadybugConfig()
-                {
-                    JsonFile = new FileSettings()
-                    {
-                        FileName = $"LBTBHoM_Materials.json",
-                        Directory = Path.GetTempPath()
-                    }
-                };
-            }
-
-            TimeSpan timeSinceLastUpdate = DateTime.Now - File.GetCreationTime(config.JsonFile.GetFullFileName());
-            if (timeSinceLastUpdate.Days > config.CacheFileMaximumAge)
-                File.Delete(config.JsonFile.GetFullFileName());
-
-            if (!File.Exists(config.JsonFile.GetFullFileName()))
-            {
-                string script = Path.Combine(Engine.LadybugTools.Query.PythonCodeDirectory(), "LadybugTools_Toolkit\\src\\ladybugtools_toolkit\\bhom\\wrapped", "get_material.py");
-
-                string cmdCommand = $"{m_environment.Executable} {script} -j \"{config.JsonFile.GetFullFileName()}\"";
-
-                Engine.Python.Compute.RunCommandStdout(command: cmdCommand, hideWindows: true);
-            }
-
-            List<object> materialObjects = Pull(new FilterRequest(), actionConfig: config).ToList();
-
-            m_executeSuccess = true;
-            return materialObjects.Where(m => (m as IEnergyMaterialOpaque).Name.Contains(command.Filter)).ToList();
-        }
-
-        /**************************************************/
-
-        private List<object> RunCommand(GetTypologyCommand command, ActionConfig actionConfig)
-        {
-            LadybugConfig config;
-
-            if (actionConfig?.GetType() == typeof(LadybugConfig))
-            {
-                config = (LadybugConfig)actionConfig;
-                config.JsonFile = new FileSettings()
-                {
-                    FileName = $"LBTBHoM_Typologies.json",
-                    Directory = Path.GetTempPath()
-                };
-            }
-            else
-            {
-                config = new LadybugConfig()
-                {
-                    JsonFile = new FileSettings()
-                    {
-                        FileName = $"LBTBHoM_Typologies.json",
-                        Directory = Path.GetTempPath()
-                    }
-                };
-            }
-
-            TimeSpan timeSinceLastUpdate = DateTime.Now - File.GetCreationTime(config.JsonFile.GetFullFileName());
-            if (timeSinceLastUpdate.Days > config.CacheFileMaximumAge)
-                File.Delete(config.JsonFile.GetFullFileName());
-
-            if (!File.Exists(config.JsonFile.GetFullFileName()))
-            {
-                string script = Path.Combine(Engine.LadybugTools.Query.PythonCodeDirectory(), "LadybugTools_Toolkit\\src\\ladybugtools_toolkit\\bhom\\wrapped", "get_typology.py");
-
-                string cmdCommand = $"{m_environment.Executable} {script} -j \"{config.JsonFile.GetFullFileName()}\"";
-
-                Engine.Python.Compute.RunCommandStdout(command: cmdCommand, hideWindows: true);
-            }
-
-            List<object> typologyObjects = Pull(new FilterRequest(), actionConfig: config).ToList();
-
-            m_executeSuccess = true;
-            return typologyObjects.Where(m => (m as Typology).Name.Contains(command.Filter)).ToList();
-        }
-
-        /**************************************************/
-
-        private List<object> RunCommand(RunSimulationCommand command, ActionConfig actionConfig)
-        {
-            // validation prior to passing to Python
-            if (command.EPWFile == null)
-            {
-                BH.Engine.Base.Compute.RecordError($"{nameof(command.EPWFile)} input cannot be null.");
-                return null;
-            }
-
-            if (!File.Exists(command.EPWFile.GetFullFileName()))
-            {
-                BH.Engine.Base.Compute.RecordError($"FIle '{command.EPWFile.GetFullFileName()}' does not exist.");
-                return null;
-            }
-
-            if (command.GroundMaterial == null)
-            {
-                BH.Engine.Base.Compute.RecordError($"{nameof(command.GroundMaterial)} input cannot be null.");
-                return null;
-            }
-
-            if (command.ShadeMaterial == null)
-            {
-                BH.Engine.Base.Compute.RecordError($"{nameof(command.ShadeMaterial)} input cannot be null.");
-                return null;
-            }
-
-            // construct adapter and config
-            LadybugConfig config = new LadybugConfig()
-            {
-                JsonFile = new FileSettings()
-                {
-                    FileName = $"LBTBHoM_{Guid.NewGuid()}.json",
-                    Directory = Path.GetTempPath()
-                }
-            };
-
-            // construct the base object and file to be passed to Python for simulation
-            SimulationResult simulationResult = new SimulationResult()
-            {
-                EpwFile = Path.GetFullPath(command.EPWFile.GetFullFileName()).Replace(@"\", "/"),
-                GroundMaterial = command.GroundMaterial,
-                ShadeMaterial = command.ShadeMaterial,
-                Name = Engine.LadybugTools.Compute.SimulationID(command.EPWFile.GetFullFileName(), command.GroundMaterial, command.ShadeMaterial)
-            };
-
-            // push object to json file
-            Push(new List<SimulationResult>() { simulationResult }, actionConfig: config);
-
-            // locate the Python file containing the simulation code
-            string script = Path.Combine(Engine.LadybugTools.Query.PythonCodeDirectory(), "LadybugTools_Toolkit\\src\\ladybugtools_toolkit\\bhom\\wrapped", "simulation_result.py");
-
-            // run the simulation
-            string cmdCommand = $"{m_environment.Executable} {script} -j \"{config.JsonFile.GetFullFileName()}\"";
-            Engine.Python.Compute.RunCommandStdout(command: cmdCommand, hideWindows: true);
-
-            // reload from Python results
-            List<object> simulationResultPopulated = Pull(new FilterRequest(), actionConfig: config).ToList();
-
-            // remove temporary file
-            File.Delete(config.JsonFile.GetFullFileName());
-
-            m_executeSuccess = true;
-            return simulationResultPopulated;
-        }
-
-        /**************************************************/
-
-        private List<object> RunCommand(RunExternalComfortCommand command, ActionConfig actionConfig)
-        {
-            if (command.SimulationResult == null)
-            {
-                BH.Engine.Base.Compute.RecordError($"{nameof(command.SimulationResult)} input cannot be null.");
-                return null;
-            }
-
-            if (command.Typology == null)
-            {
-                BH.Engine.Base.Compute.RecordError($"{nameof(command.Typology)} input cannot be null.");
-                return null;
-            }
-
-            LadybugConfig config = new LadybugConfig()
-            {
-                JsonFile = new FileSettings()
-                {
-                    FileName = $"LBTBHoM_{Guid.NewGuid()}.json",
-                    Directory = Path.GetTempPath()
-                }
-            };
-
-            // construct the base object
-            ExternalComfort externalComfort = new ExternalComfort()
-            {
-                SimulationResult = command.SimulationResult,
-                Typology = command.Typology,
-            };
-
-            // push objects to json file
-            Push(new List<ExternalComfort>() { externalComfort }, actionConfig: config);
-
-            // locate the Python file containing the simulation code
-            string script = Path.Combine(Engine.LadybugTools.Query.PythonCodeDirectory(), "LadybugTools_Toolkit\\src\\ladybugtools_toolkit\\bhom\\wrapped", "external_comfort.py");
-
-            // run the calculation
-            string cmdCommand = $"{m_environment.Executable} {script} -j \"{config.JsonFile.GetFullFileName()}\"";
-            Engine.Python.Compute.RunCommandStdout(command: cmdCommand, hideWindows: true);
-
-            // reload from Python results
-            List<object> externalComfortPopulated = Pull(new FilterRequest(), actionConfig: config).ToList();
-
-            // remove temporary file
-            File.Delete(config.JsonFile.GetFullFileName());
-
-            m_executeSuccess = true;
-            return externalComfortPopulated;
-        }
-
-        /**************************************************/
-
-        private List<object> RunCommand(HeatPlotCommand command, ActionConfig actionConfig)
-        {
-            if (command.EPWFile == null)
-            {
-                BH.Engine.Base.Compute.RecordError($"{nameof(command.EPWFile)} input cannot be null.");
-                return null;
-            }
-
-            if (!System.IO.File.Exists(command.EPWFile.GetFullFileName()))
-            {
-                BH.Engine.Base.Compute.RecordError($"File '{command.EPWFile}' does not exist.");
-                return null;
-            }
-
-            string epwFile = System.IO.Path.GetFullPath(command.EPWFile.GetFullFileName());
-
-            string script = Path.Combine(Engine.LadybugTools.Query.PythonCodeDirectory(), "LadybugTools_Toolkit\\src\\ladybugtools_toolkit\\bhom\\wrapped\\plot", "heatmap.py");
-
-            //check if the colourmap is valid for user warning, but run with input anyway as the map could be defined separately.
-            string colourMap = command.ColourMap;
-            if (colourMap.ColourMapValidity())
-                colourMap = colourMap.ToColourMap().FromColourMap();
-
-            string returnFile = Path.GetTempFileName();
-
-            // run the process
-            string cmdCommand = $"{m_environment.Executable} {script} -e \"{epwFile}\" -dtk \"{command.EPWKey.ToText()}\" -cmap \"{colourMap}\" -r \"{returnFile.Replace('\\', '/')}\" -p \"{command.OutputLocation}\"";
-            string result = Engine.Python.Compute.RunCommandStdout(command: cmdCommand, hideWindows: true);
-
-            if (!File.Exists(result))
-            {
-                BH.Engine.Base.Compute.RecordError($"An error occurred while running the command: {result}");
-                File.Delete(returnFile);
-                return new List<object>();
-            }
-
-            CustomObject obj = (CustomObject)BH.Engine.Serialiser.Convert.FromJson(System.IO.File.ReadAllText(returnFile));
-            File.Delete(returnFile);
-            PlotInformation info = Convert.ToPlotInformation(obj, new CollectionData());
-
-            m_executeSuccess = true;
-            return new List<object>() { info };
-        }
-
-        /**************************************************/
-
-        private List<object> RunCommand(WindroseCommand command, ActionConfig actionConfig)
-        {
-            if (command.EPWFile == null)
-            {
-                BH.Engine.Base.Compute.RecordError($"{nameof(command.EPWFile)} input cannot be null.");
-                return null;
-            }
-
-            if (!System.IO.File.Exists(command.EPWFile.GetFullFileName()))
-            {
-                BH.Engine.Base.Compute.RecordError($"File '{command.EPWFile.GetFullFileName()}' does not exist.");
-                return null;
-            }
-
-            if (command.AnalysisPeriod == null)
-            {
-                BH.Engine.Base.Compute.RecordError($"Please input a valid analysis period to run this command.");
-                return null;
-            }
-
-            string epwFile = System.IO.Path.GetFullPath(command.EPWFile.GetFullFileName());
-
-            string script = Path.Combine(Engine.LadybugTools.Query.PythonCodeDirectory(), "LadybugTools_Toolkit\\src\\ladybugtools_toolkit\\bhom\\wrapped\\plot", "windrose.py");
-
-            //check if the colourmap is valid for user warning, but run with input anyway as the map could be defined separately.
-            string colourMap = command.ColourMap;
-            if (colourMap.ColourMapValidity())
-                colourMap = colourMap.ToColourMap().FromColourMap();
-
-            string returnFile = Path.GetTempFileName();
-
-            // run the process
-            string cmdCommand = $"{m_environment.Executable} {script} -e \"{epwFile}\" -ap \"{command.AnalysisPeriod.FromBHoM().Replace("\"", "\\\"")}\" -cmap \"{colourMap}\" -bins \"{command.NumberOfDirectionBins}\" -r \"{returnFile.Replace('\\', '/')}\" -p \"{command.OutputLocation}\"";
-            string result = Engine.Python.Compute.RunCommandStdout(command: cmdCommand, hideWindows: true);
-
-            if (!File.Exists(result))
-            {
-                BH.Engine.Base.Compute.RecordError($"An error occurred while running the command: {result}");
-                File.Delete(returnFile);
-                return new List<object>();
-            }
-
-            CustomObject obj = (CustomObject)BH.Engine.Serialiser.Convert.FromJson(System.IO.File.ReadAllText(returnFile));
-            File.Delete(returnFile);
-            PlotInformation info = Convert.ToPlotInformation(obj, new WindroseData());
-
-            m_executeSuccess = true;
-            return new List<object> { info };
-        }
-
-        /**************************************************/
-
-        private List<object> RunCommand(UTCIHeatPlotCommand command, ActionConfig actionConfig)
-        {
-            if (command.EPWFile == null)
-            {
-                BH.Engine.Base.Compute.RecordError($"{nameof(command.EPWFile)} input cannot be null.");
-                return null;
-            }
-
-            if (!System.IO.File.Exists(command.EPWFile.GetFullFileName()))
-            {
-                BH.Engine.Base.Compute.RecordError($"File '{command.EPWFile.GetFullFileName()}' does not exist.");
-                return null;
-            }
-
-            if (command.GroundMaterial == null)
-            {
-                BH.Engine.Base.Compute.RecordError($"Please input a valid ground material to run this command.");
-                return null;
-            }
-
-            if (command.ShadeMaterial == null)
-            {
-                BH.Engine.Base.Compute.RecordError($"Please input a valid shade material to run this command.");
-                return null;
-            }
-
-            if (command.Typology == null)
-            {
-                BH.Engine.Base.Compute.RecordError($"Please input a valid Typology to run this command.");
-            }
-
-            if (!(command.BinColours.Count == 10 || command.BinColours.Count == 0))
-            {
-                BH.Engine.Base.Compute.RecordError($"When overriding bin colours 10 colours must be provided, but {command.BinColours.Count} colours were provided instead.");
-                return null;
-            }
-
-            List<string> colours = command.BinColours.Select(x => x.ToHexCode()).ToList();
-
-            string hexColours = $"[\"{string.Join("\",\"", colours)}\"]";
-            if (hexColours == "[\"\"]")
-                hexColours = "[]";
-
-            Dictionary<string, string> inputObjects = new Dictionary<string, string>()
-            {
-                { "ground_material",  command.GroundMaterial.FromBHoM() },
-                { "shade_material", command.ShadeMaterial.FromBHoM() },
-                { "typology", command.Typology.FromBHoM() },
-                { "bin_colours", hexColours }
-            };
-
-            string argFile = Path.GetTempFileName();
-            File.WriteAllText(argFile, inputObjects.ToJson());
-
-            string epwFile = System.IO.Path.GetFullPath(command.EPWFile.GetFullFileName());
-
-            string script = Path.Combine(Engine.LadybugTools.Query.PythonCodeDirectory(), "LadybugTools_Toolkit\\src\\ladybugtools_toolkit\\bhom\\wrapped\\plot", "utci_heatmap.py");
-
-            string returnFile = Path.GetTempFileName();
-
-            // run the process
-            string cmdCommand = $"{m_environment.Executable} \"{script}\" -e \"{epwFile}\" -in \"{argFile}\" -ws \"{command.WindSpeedMultiplier}\" -r \"{returnFile.Replace('\\', '/')}\" -sp \"{command.OutputLocation}\"";
-            string result = Engine.Python.Compute.RunCommandStdout(command: cmdCommand, hideWindows: true);
-
-            string resultFile = result.Split('\n').Last();
-
-            if (!File.Exists(resultFile))
-            {
-                BH.Engine.Base.Compute.RecordError($"An error occurred while running the command: {result}");
-                File.Delete(returnFile);
-                File.Delete(argFile);
-                return new List<object>();
-            }
-
-            CustomObject obj = (CustomObject)BH.Engine.Serialiser.Convert.FromJson(System.IO.File.ReadAllText(returnFile));
-            File.Delete(returnFile);
-            File.Delete(argFile);
-            PlotInformation info = Convert.ToPlotInformation(obj, new UTCIData());
-
-            m_executeSuccess = true;
-            return new List<object> { info };
-        }
-
-        /**************************************************/
-
-        private List<object> RunCommand(DiurnalPlotCommand command, ActionConfig actionConfig)
-        {
-            if (command.EPWFile == null)
-            {
-                BH.Engine.Base.Compute.RecordError($"{nameof(command.EPWFile)} input cannot be null.");
-                return null;
-            }
-
-            if (!System.IO.File.Exists(command.EPWFile.GetFullFileName()))
-            {
-                BH.Engine.Base.Compute.RecordError($"File '{command.EPWFile}' does not exist.");
-                return null;
-            }
-
-            if (command.Period == DiurnalPeriod.Undefined)
-            {
-                BH.Engine.Base.Compute.RecordError("Please provide a valid diurnal period.");
-                return null;
-            }
-
-            if (command.EPWKey == EPWKey.Undefined)
-            {
-                BH.Engine.Base.Compute.RecordError("Please provide a valid EPW key.");
-                return null;
-            }
-
-            command.Title = command.Title.SanitiseString();
-
-            string epwFile = System.IO.Path.GetFullPath(command.EPWFile.GetFullFileName());
-
-            string script = Path.Combine(Engine.LadybugTools.Query.PythonCodeDirectory(), "LadybugTools_Toolkit\\src\\ladybugtools_toolkit\\bhom\\wrapped\\plot", "diurnal.py");
-
-            string returnFile = Path.GetTempFileName();
-
-            // run the process
-            string cmdCommand = $"{m_environment.Executable} {script} -e \"{epwFile}\" -dtk \"{command.EPWKey.ToText()}\" -c \"{command.Colour.ToHexCode()}\" -t \"{command.Title}\" -ap \"{command.Period.ToString().ToLower()}\" -r \"{returnFile.Replace('\\', '/')}\" -p \"{command.OutputLocation}\"";
-            string result = Engine.Python.Compute.RunCommandStdout(command: cmdCommand, hideWindows: true);
-
-            string resultFile = result.Split('\n').Last();
-
-            if (!File.Exists(resultFile))
-            {
-                BH.Engine.Base.Compute.RecordError($"An error occurred while running the command: {result}");
-                File.Delete(returnFile);
-                return new List<object>();
-            }
-
-            CustomObject obj = (CustomObject)BH.Engine.Serialiser.Convert.FromJson(System.IO.File.ReadAllText(returnFile));
-            File.Delete(returnFile);
-            PlotInformation info = Convert.ToPlotInformation(obj, new CollectionData());
-
-            m_executeSuccess = true;
-            return new List<object>() { info };
-        }
-
-        /**************************************************/
-
-        private List<object> RunCommand(SunPathPlotCommand command, ActionConfig actionConfig)
-        {
-            if (command.EPWFile == null)
-            {
-                BH.Engine.Base.Compute.RecordError($"{nameof(command.EPWFile)} input cannot be null.");
-                return null;
-            }
-            
-            if (!System.IO.File.Exists(command.EPWFile.GetFullFileName()))
-            {
-                BH.Engine.Base.Compute.RecordError($"File '{command.EPWFile.GetFullFileName()}' does not exist.");
-                return null;
-            }
-
-            if (command.AnalysisPeriod == null)
-            {
-                BH.Engine.Base.Compute.RecordError($"{nameof(command.EPWFile)} input cannot be null.");
-                return null;
-            }
-
-            if (command.SunSize < 1)
-            {
-                BH.Engine.Base.Compute.RecordError($"Sun size must be greater than or equal to 1.");
-                return null;
-            }
-
-            string epwFile = System.IO.Path.GetFullPath(command.EPWFile.GetFullFileName());
-
-            string script = Path.Combine(Engine.LadybugTools.Query.PythonCodeDirectory(), "LadybugTools_Toolkit\\src\\ladybugtools_toolkit\\bhom\\wrapped\\plot", "sunpath.py");
-
-            string returnFile = Path.GetTempFileName();
-
-            //run the process
-            string cmdCommand = $"{m_environment.Executable} {script} -e \"{epwFile}\" -s {command.SunSize} -ap \"{command.AnalysisPeriod.FromBHoM().Replace("\"", "\\\"")}\" -r \"{returnFile.Replace('\\', '/')}\" -p \"{command.OutputLocation}\"";
-            string result = Engine.Python.Compute.RunCommandStdout(cmdCommand, hideWindows: true);
-
-            string resultFile = result.Split('\n').Last();
-
-            if (!File.Exists(resultFile))
-            {
-                BH.Engine.Base.Compute.RecordError($"An error occurred while running the command: {result}");
-                File.Delete(returnFile);
-                return new List<object>();
-            }
-
-            CustomObject obj = (CustomObject)BH.Engine.Serialiser.Convert.FromJson(System.IO.File.ReadAllText(returnFile));
-            File.Delete(returnFile);
-            PlotInformation info = Convert.ToPlotInformation(obj, new SunPathData());
-
-            m_executeSuccess = true;
-            return new List<object>() { info };
-        }
-
-        /**************************************************/
-
-        private List<object> RunCommand(SolarRadiationPlotCommand command, ActionConfig actionConfig)
-        {
-            if (command.EPWFile == null)
-            {
-                BH.Engine.Base.Compute.RecordError($"{nameof(command.EPWFile)} input cannot be null.");
-                return null;
-            }
-
-            if (!System.IO.File.Exists(command.EPWFile.GetFullFileName()))
-            {
-                BH.Engine.Base.Compute.RecordError($"File '{command.EPWFile.GetFullFileName()}' does not exist.");
-                return null;
-            }
-
-            if (command.Directions < 3)
-            {
-                BH.Engine.Base.Compute.RecordError($"Azimuths must be greater than or equal to 1.");
-                return null;
-            }
-
-            if (command.Tilts < 3)
-            {
-                BH.Engine.Base.Compute.RecordError($"Altitudes must be greater than or equal to 1");
-                return null;
-            }
-
-            if (command.IrradianceType == IrradianceType.Undefined)
-            {
-                BH.Engine.Base.Compute.RecordError($"Please provide a valid Irradiance Type.");
-                return null;
-            }
-
-            if (command.AnalysisPeriod == null)
-            {
-                BH.Engine.Base.Compute.RecordError($"{nameof(AnalysisPeriod)} input cannot be null.");
-                return null;
-            }
-
-            //check if the colourmap is valid for user warning, but run with input anyway as the map could be defined separately.
-            string colourMap = command.ColourMap;
-            if (colourMap.ColourMapValidity())
-                colourMap = colourMap.ToColourMap().FromColourMap();
-
-            string epwFile = System.IO.Path.GetFullPath(command.EPWFile.GetFullFileName());
-            string returnFile = Path.GetTempFileName();
-
-            string script = Path.Combine(Engine.LadybugTools.Query.PythonCodeDirectory(), "LadybugTools_Toolkit\\src\\ladybugtools_toolkit\\bhom\\wrapped\\plot", "directional_solar_radiation.py");
-
-            string cmdCommand = $"{m_environment.Executable} {script} -e \"{epwFile}\" -d {command.Directions} -ti {command.Tilts} -ir {command.IrradianceType} -cmap \"{colourMap}\" -t \"{command.Title}\" -ap \"{command.AnalysisPeriod.FromBHoM().Replace("\"", "\\\"")}\" -p \"{command.OutputLocation}\" -r \"{returnFile.Replace('\\', '/')}\"";
-            string result = Engine.Python.Compute.RunCommandStdout(cmdCommand, hideWindows: true);
-
-            string resultFile = result.Split('\n').Last();
-
-            if (!File.Exists(resultFile))
-            {
-                BH.Engine.Base.Compute.RecordError($"An error occurred while running the command: {result}");
-                File.Delete(returnFile);
-                return new List<object>();
-            }
-
-            CustomObject obj = (CustomObject)BH.Engine.Serialiser.Convert.FromJson(System.IO.File.ReadAllText(returnFile));
-            File.Delete(returnFile);
-            PlotInformation info = Convert.ToPlotInformation(obj, new SolarRadiationData());
-
-            m_executeSuccess = true;
-            return new List<object>() { info };
         }
 
         /**************************************************/
